# openapi.yaml
openapi: 3.0.3
info:
  title: Watch Merchant API
  description: |
    Demo watch merchant API with static product catalog for AP2 integration.

    This API provides endpoints for browsing luxury watches, managing shopping carts,
    calculating pricing, checking inventory, and placing orders.
  version: 1.0.0
  contact:
    name: API Support
    email: support@watchmerchant.demo
servers:
  - url: http://localhost:8765/api/v1
    description: Local development server
  - url: https://api.watchmerchant.demo/v1
    description: Production server
...
paths:
  /products:
    get:
      tags: [Products]
      summary: List products
      description: Get a list of watches with optional filtering, sorting, and pagination
      operationId: watches_listProducts
      parameters:
        - name: search
          in: query
          description: Search term for name, brand, or description
          schema:
            type: string
        - name: brand
          in: query
          description: Filter by brand (can be multiple)
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
        - name: category
          in: query
          description: Filter by category (can be multiple)
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
        - name: priceMin
          in: query
          description: Minimum price filter
          schema:
            type: number
        - name: priceMax
          in: query
          description: Maximum price filter
          schema:
            type: number
        - name: minRating
          in: query
          description: Minimum rating filter
          schema:
            type: number
            minimum: 0
            maximum: 5
        - name: tags
          in: query
          description: Filter by tags
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
        - name: inStock
          in: query
          description: Only show in-stock items
          schema:
            type: boolean
        - name: sort
          in: query
          description: Sort order
          schema:
            type: string
            enum: [price_asc, price_desc, rating, newest]
        - name: page
          in: query
          description: Page number (default 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page (default 20, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  meta:
                    type: object
                    properties:
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      description: Retrieve detailed information about a specific watch
      operationId: watches_getProduct
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/slug/{slug}:
    get:
      tags: [Products]
      summary: Get product by slug
      description: Retrieve product by URL-friendly slug
      operationId: watches_getProductBySlug
      parameters:
        - name: slug
          in: path
          required: true
          description: Product slug
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        name:
          type: string
        slug:
          type: string
        brand:
          type: string
        price:
          type: number
        currency:
          type: string
        images:
          type: array
          items:
            type: string
        description:
          type: string
        shortDescription:
          type: string
        stock:
          type: integer
        categories:
          type: array
          items:
            type: string
        specs:
          type: object
          properties:
            caseSize:
              type: string
            caseMaterial:
              type: string
            strapMaterial:
              type: string
            movement:
              type: string
            waterResistance:
              type: string
            crystalType:
              type: string
            warranty:
              type: string
        rating:
          type: number
        reviewCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        weight:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object